Move MTDTopology to MTDGeometryBuilder, add methods for BTL topology
