+++ Geometry/MTDGeometryBuilder/plugins/MTDTopologyEP.cc
@@ -6,7 +6,7 @@
 #include "FWCore/Framework/interface/ModuleFactory.h"
 #include "FWCore/Framework/interface/ESHandle.h"
 #include "FWCore/Framework/interface/ESProducer.h"
-#include "Geometry/MTDNumberingBuilder/interface/MTDTopology.h"
+#include "Geometry/MTDGeometryBuilder/interface/MTDTopology.h"
 #include "Geometry/MTDCommonData/interface/MTDTopologyMode.h"
 #include "Geometry/Records/interface/MTDTopologyRcd.h"
 #include "CondFormats/GeometryObjects/interface/PMTDParameters.h"
