+++ Geometry/MTDGeometryBuilder/test/DD4hep_TestPixelTopology.cc
@@ -33,7 +33,7 @@
 #include "DataFormats/GeometrySurface/interface/RectangularPlaneBounds.h"
 
 #include "Geometry/Records/interface/MTDTopologyRcd.h"
-#include "Geometry/MTDNumberingBuilder/interface/MTDTopology.h"
+#include "Geometry/MTDGeometryBuilder/interface/MTDTopology.h"
 #include "Geometry/MTDGeometryBuilder/interface/MTDGeometry.h"
 #include "Geometry/MTDGeometryBuilder/interface/MTDGeomDetUnit.h"
 #include "Geometry/Records/interface/MTDDigiGeometryRecord.h"
