+++ Geometry/MTDNumberingBuilder/python/mtdTopology_cfi.py
@@ -1,0 +1,1 @@
+from Geometry.MTDGeometryBuilder.mtdTopology_cfi import *
