+++ RecoLocalFastTime/FTLRecProducers/plugins/MTDTimeCalibESProducer.cc
@@ -4,7 +4,7 @@
 #include "Geometry/Records/interface/MTDTopologyRcd.h"
 
 #include "Geometry/MTDGeometryBuilder/interface/MTDGeometry.h"
-#include "Geometry/MTDNumberingBuilder/interface/MTDTopology.h"
+#include "Geometry/MTDGeometryBuilder/interface/MTDTopology.h"
 
 #include "RecoLocalFastTime/FTLCommonAlgos/interface/MTDTimeCalib.h"
