+++ RecoMTD/DetLayers/plugins/MTDDetLayerGeometryESProducer.cc
@@ -16,7 +16,7 @@
 #include "Geometry/Records/interface/MTDDigiGeometryRecord.h"
 #include "Geometry/MTDGeometryBuilder/interface/MTDGeometry.h"
 #include "Geometry/Records/interface/MTDTopologyRcd.h"
-#include "Geometry/MTDNumberingBuilder/interface/MTDTopology.h"
+#include "Geometry/MTDGeometryBuilder/interface/MTDTopology.h"
 
 #include "RecoMTD/DetLayers/interface/ETLDetLayerGeometryBuilder.h"
 #include "RecoMTD/DetLayers/interface/BTLDetLayerGeometryBuilder.h"
