+++ RecoTracker/GeometryESProducer/plugins/TrackerMTDRecoGeometryESProducer.cc
@@ -6,7 +6,7 @@
 #include "DataFormats/TrackerCommon/interface/TrackerTopology.h"
 #include "Geometry/Records/interface/TrackerTopologyRcd.h"
 #include "Geometry/MTDGeometryBuilder/interface/MTDGeometry.h"
-#include "Geometry/MTDNumberingBuilder/interface/MTDTopology.h"
+#include "Geometry/MTDGeometryBuilder/interface/MTDTopology.h"
 #include "Geometry/Records/interface/MTDTopologyRcd.h"
 #include "Geometry/Records/interface/MTDDigiGeometryRecord.h"
